added getnode function
